refactor: changed the structure of file -- project
