+++ src/components/pages/ads-page/BtnComponent.jsx
