+++ src/components/pages/ads-page/NumeriComponent.jsx
