+++ src/components/pages/ads-page/StringComponent.jsx
