+++ src/components/pages/ads-page/SwitchComponent.jsx
