+++ src/components/pages/share/LoadingText.jsx
